requirements.txt
google-genai==0.3.0
pydantic==2.5.3
python-dotenv
